+++ site/src/components/WorkspaceDeletion/utils.ts
@@ -11,7 +11,7 @@ const IMPENDING_DELETION_DISPLAY_THRESHOLD = 14; // 14 days
  * @param {TypesGen.Workspace} workspace
  * @returns {boolean}
  */
-export const displayImpendingDeletion = (
+export const displayDormantDeletion = (
   workspace: Workspace,
   allowAdvancedScheduling: boolean,
   allowWorkspaceActions: boolean,
