+++ site/src/pages/WorkspacePage/Workspace.tsx
@@ -265,7 +265,7 @@ export const Workspace: FC<React.PropsWithChildren<WorkspaceProps>> = ({
               />
             </Cond>
             <Cond>
-              {/* <ImpendingDeletionBanner/> determines its own visibility */}
+              {/* <DormantWorkspaceBanner/> determines its own visibility */}
               <DormantWorkspaceBanner
                 workspaces={[workspace]}
                 shouldRedisplayBanner={
